Display an alert when a new connection is not answered
When the computer is on but you are away, you may miss some prompt
dialogs to allow or deny a connection. This way we warn the user about
something that happened, in order to review the UI.

Further work should maybe count and display the number of a connection,
and perhaps display a max of 5 different events.

It should also desirable to display a small number over the app icon.