UI: added dialog to add/modify rules from the UI
Allow to add or modify rules from the UI.

The rules of all nodes are listed in the tab Rules.
Clicking on the name of the rule will display the connections that
matched the rule.

In this view 2 buttons are displayed: edit rule and delete rule.

Editing a rule will edit the rule for a particular node. If you want
the changes to be applied to all nodes you have to select that option.

All fields can have a regular expression. If the regular expression is
not valid, a message will be displayed on the dialog.

If the rule is applied correctly, "Rule applied" will be shown.
Otherwise an error will appear.

Bear in mind that this still a WIP commit. Bugs and errors will
appear that we'll have to fix.