ui, prompt: improved showing app path and args
The pop-ups display 3 labels:
 - The name of the app (Firefox, aMule, Chromium, etc)
 - The path of the binary (/usr/bin/telnet)
 - The cmd line that was typed or executed (telnet 1.1.1.1)

The app name is always displayed. If we haven't found the app for a
given connection, "Unknown process" is displayed".

The app path is where the binary is located, and the app arguments is
the cmd line that wsa typed by the user, or executed by a process.

So for example, if you execute: telnet 1.1.1.1, the user wouldn't know
what is the path of the application. Someone could have faked it, thus
it's useful to display path to the binary.

On the other hand, if you execute /usr/bin/telnet.netkit 1.1.1.1, the
binary path is already part of the cmdline, so there's no need to show
that information to the user.