ui, pop-ups: added app description, better icon discovery
Added the description of an app to the pop-ups, to help users know
what an application is or does.

The discovery of app icons has been improved for those edge cases where
the system is not properly configured and we were not able to get the
icon of the app.