ui, events: better rows selection
As events are received from the daemon, the list of events is reloaded,
and the selection of rows was lost.
This caused several side effects: difficulty to delete multiple rules,
as well as analyze listed events because they were reloaded constantly.

Now if a row is selected, the new events are not loaded until rows
selection is not cleared.

closes #560