ui: make GUI usable with thousands of records
The GUI was not usable when there were thousands of records. It used to
get frozen while loading large amount of data, specially when using a DB
stored on disk. Besides, when entering in the details view of an event
it was not possible to filter records by text.

Changes:
- Load by default 50 records instead of all records returned by a query.
- Improved/fixed detail views' queries.

New:
- Allow to filter data in all views, including detail views.
- Allow to limit records in all views.
- Added a help button to display a ballon message with useful
  information.

Fixes:
- selection jumping continuously to row 256 when there were more than
  256 rows.
https://github.com/evilsocket/opensnitch/issues/565#issuecomment-997346990

Even without setting query limits, the GUI is more usable than before,
but it needs more work. WIP.