ui, events: fixed delay displaying the main window
If the events were saved to a DB on disk,
there were long delays when opening the events window.

Now the events window loads faster, but the Events view is not as
performant as before. It means that while you're monitoring connections,
the GUI may consume 1-5% of the CPU and ~25% when you apply a filter
(only in that moment).

With the previous approach the CPU usage was close to 0-1% in both
cases, so we need to find a way of not causing long delays while being
performant.

More about this:
https://github.com/evilsocket/opensnitch/issues/566#issuecomment-1014930627