ui,rules: improved regular expressions
Up until now we detected automatically if a rule was a regular
expression. But if the path or cmdline of a process had certain
characters (backslash), it was not easy to autodetect it, because the
backslash in particular is sent and received encoded (\\).

With this change the user will have to mark manually if the
path/cmdline of a process is a regular expression or no.

Closes #706