pop-ups: filter by process path if cmdline is relative
It's a good practice to filter by command line + process path (and
others parameters), to limit what an application can do.

Thus, if the user selects "from this command line", and the command
launched is not an absolute path to the binary, we'll create a rule
to filter by process path + process cmdline.