db: use shared cache for in-memory db
Due to how QsqlDatabase works from different threads, a connection to
the DB can only be used from the thread that created it.

In order to access a database from different threads you create a new
QsqlDatabase connection to the DB specifying the DB name/file.

With DBs written to the disk, the database file/name is always the
same (say /tmp/file.db).

But with in-memory databases, whenever you create a new connection with
:memory:, it always create a new in-memory DB, so the DB is empty for
the Cleaner() task, and that's why it was not deleting old events.

Closes #844