deserialize rules operator list correctly
In 582b91e7856179d7d7adbb7083eb5e4764508ee8 we disabled sending/parsing
list operators as JSON strings. Instead, now it's sent/parsed as
protobuf Rule, and saved to disk as JSON array, which ease the task of
manually creating new rules if needed.

This change was missing in the previous commit.