pkgs,rpm: added python311-PyQt6 as dependency
Some distros like OpenSuse Leap 15.6 ships with python3.6 by default.
PyQt6 is not available for python3.6, but python311-PyQt6 is.
So there's no generic python3-PyQt6, but python311-PyQt6.

In this specific case, the user will have to change the default python3
version to python3.11. And they'll have to install some packages like
slugify from pip.

But as a best effort, we'll try to install PyQt6 from the distro
repositories.